TDW-2277 created dependencies.js file
